Add files via upload
During this project, we made a program that used user stories and an agile approach to help us finish the program. The user stories helped us stay on track and finish things promptly and not waste time on unnecessary features. We also had a hurdle we had to change pretty far into the project, and using user stories helped us accomplish that without having to rewrite the whole code. I intend to use user stories and the agile approach as opposed to the waterfall approach in my future works because that's what seemed to work best here to ensure we could adapt to changes and requests by the customer.